refactor: security 구조 변경 (#24)
