+++ src/main/java/project/backend/business/auth/oauth/KakaoUserDetailsService.java
@@ -8,8 +8,8 @@ import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserServ
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import project.backend.common.auth.oauth.KakaoUserDetails;
-import project.backend.common.auth.oauth.KakaoUserInfo;
+import project.backend.security.oauth.KakaoUserDetails;
+import project.backend.security.oauth.KakaoUserInfo;
 import project.backend.entity.user.User;
 import project.backend.repository.user.UserRepository;
