+++ src/main/java/project/backend/presentation/post/controller/PostController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.*;
 import project.backend.business.post.PostService;
 import project.backend.business.post.dto.PostDetailDto;
 import project.backend.business.post.dto.PostListDto;
-import project.backend.common.auth.aop.AssignCurrentUserInfo;
-import project.backend.common.auth.aop.AssignOrNullCurrentUserInfo;
-import project.backend.common.auth.aop.CurrentUserInfo;
+import project.backend.security.aop.AssignCurrentUserInfo;
+import project.backend.security.aop.AssignOrNullCurrentUserInfo;
+import project.backend.security.aop.CurrentUserInfo;
 import project.backend.presentation.post.dto.request.SummaryUrlRequest;
 import project.backend.presentation.post.dto.request.UpdatePostRequest;
 import project.backend.presentation.post.dto.response.CreateUpdatePostResponse;
