+++ src/main/java/project/backend/presentation/user/controller/UserController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import project.backend.business.user.UserService;
-import project.backend.common.auth.aop.AssignCurrentUserInfo;
-import project.backend.common.auth.aop.CurrentUserInfo;
+import project.backend.security.aop.AssignCurrentUserInfo;
+import project.backend.security.aop.CurrentUserInfo;
 import project.backend.entity.user.User;
 
 @RestController
