+++ src/main/java/project/backend/security/aop/AssignCurrentUserInfo.java
@@ -1,4 +1,4 @@
-package project.backend.common.auth.aop;
+package project.backend.security.aop;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,6 @@ import java.lang.annotation.Target;
 
 @Retention(value = RetentionPolicy.RUNTIME)
 @Target(value = ElementType.METHOD)
-public @interface AssignOrNullCurrentUserInfo {
+public @interface AssignCurrentUserInfo {
+
 }
