+++ src/main/java/project/backend/security/jwt/JwtAuthenticationFailEntryPoint.java
@@ -1,4 +1,4 @@
-package project.backend.common.auth.jwt;
+package project.backend.security.jwt;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
