+++ src/main/java/project/backend/security/token/RefreshToken.java
@@ -1,4 +1,4 @@
-package project.backend.common.auth.token;
+package project.backend.security.token;
 
 import java.util.Collection;
 import org.springframework.data.annotation.Id;
